refactor: 오타 수정 및 디렉토리 변경
