+++ src/main/java/FinTo/domain/member/service/MemberServiceImpl.java
@@ -1,5 +1,6 @@
-package FinTo.domain.member;
+package FinTo.domain.member.service;
 
+import FinTo.domain.member.repository.MemberRepository;
 import FinTo.domain.member.domain.Member;
 import FinTo.domain.member.exception.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
