+++ src/main/java/FinTo/global/security/userdetails/CustomUserDetailsService.java
@@ -1,7 +1,6 @@
 package FinTo.global.security.userdetails;
 
-import FinTo.domain.member.MemberRepository;
-import FinTo.domain.member.MemberService;
+import FinTo.domain.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
